+++ src/main/java/edu/stefano/forti/pad/hashtomin/HashToMin.java
@@ -8,6 +8,7 @@ package edu.stefano.forti.pad.hashtomin;
 import java.io.IOException;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -21,6 +22,9 @@ public class HashToMin {
 
     /**
      * @param args the command line arguments
+     * @throws java.io.IOException
+     * @throws java.lang.InterruptedException
+     * @throws java.lang.ClassNotFoundException
      */
     public static void main(String[] args) throws IOException, InterruptedException, ClassNotFoundException {
         
@@ -33,8 +37,8 @@ public class HashToMin {
         
         job.setMapperClass(EdgesToAdjacencyListMapper.class);
         job.setReducerClass(EdgesToAdjacencyListReducer.class);
-        
-        job.setOutputKeyClass(IntWritable.class);
+
+        job.setOutputKeyClass(LongWritable.class);
         job.setOutputValueClass(Text.class);
         
         System.exit(job.waitForCompletion(true)? 0 : 1);
